RCI(decomp)	R Documentation

A function to calculate the relative concentration index

Description

This function calculates the relative concentration index due to Kakwani et al. It returns an object of class hci for which confidence intervals, summaries and plots are defined.

Usage


RCI(x, y, wt = NULL)
concentrationIndex(object)
## S3 method for class 'hci':
plot(x, ...)
## S3 method for class 'hci':
print(x, ...)
## S3 method for class 'hci':
summary(object, ...)
## S3 method for class 'hci':
confint(object, parm = NULL, level = 0.95, ...)
varianceConcentrationIndex(object)

Arguments

x 	 For RCI, a numeric variable containing the health measure. For the print and plot methods: an object of class hci.

y 	 a numeric variable containing the wealth measure.

wt 	 a numeric variable containing the sample weights. Defaults to NULL, which is equal to a constant sample weight for all cases.

object 	 an object of class hci - the result of a call to RCI.

parm 	 always set to NULL; provided for consistency with the generic confint.

level 	 confidence level at which confidence intervals should be calculated. Defaults to 0.95.

... 	 arguments to pass on to plot, print and summary.

Details

RCI returns the relative concentration index due to Kakwani et al in an object of class hci. A number of extractor functions have been defined. The plot function returns a plot of the Lorenz curve. The relative concentration index is defined as twice the surface between this curve and the diagonal. RCI does not check whether the health measure has the same sign for all cases. If this is not true the relative concentration index is not bounded between [-1,+1]. It is up to the user to handle this, e.g. using the correctSign function.

Value

For RCI, an object of class hci; this is a list containing the following components:

concentrationIndex 	 relative concentration index

varC 	 variance of the concentration index

fractionalRank 	 weighted fractional rank for each case

outcome 	 health measure for each case

n 	 number of cases

call 	 the original call to RCI

For plot.hci, a concentration plot.

Author(s)

Peter Konings and Niko Speybroeck

References

Kakwani et al.

See Also

correctSign

Examples


data(kenya)
example <- RCI(x = kenya$wealth, y = kenya$stunting)
plot(example)
summary(example)

[Package decomp version 0.3 Index]