rank.wt(decomp)	R Documentation

Ranks a numeric variable taking weights into account. Mimics a Stata function

Description

This function calculates the weighted fractional ranks of a numeric variable.

Usage


rank.wt(x, wt)

Arguments

x 	 numeric vector

wt 	 weights

Value

a numeric vector containing weighted fractional ranks of the elements of x.

Author(s)

Peter Konings

References

Kakwani et al., 1997.

Examples


x <- sample(1:10, size = 10, replace = TRUE)
x.weight <- seq(0, 1, length.out = 10)
rank.wt(x, wt = x.weight)

[Package decomp version 0.3 Index]